Fix TypeResolver test for Maps.newEnumMap and the invalid special treatment the wrongly-written test was getting in TypeResolver.
Disable test that asserts failure when List<List<?>> is provided as the
argument for a parameter of type List<List<T>>.  We still don't have a
good understanding of why that's supposed to fail and other similar code
is not supposed to fail -- we don't understand the underlying rule that
javac is enforcing there.

Simplify getParameterizedSupertype logic.